Move filewriting caching operation into download cache
